+++ src/driver/session.ml
@@ -1,7 +1,16 @@
-type t = { sources : Common.Source_manager.t }
+type t = {
+  sources : Common.Source_manager.t;
+  definitions : Frontend.Definition.Environment.t;
+}
 
-let create () = { sources = Common.Source_manager.create () }
+let create () =
+  {
+    sources = Common.Source_manager.create ();
+    definitions = Frontend.Definition.Environment.create ();
+  }
+
 let sources session = session.sources
+let definitions session = session.definitions
 
 let add_source session ~path ~contents =
   Common.Source_manager.add_string session.sources ~path ~contents